extcon: arizona: Move measured hp_impedance to arizona
The switch driver was already updated, however the extcon driver is not.
This means that whilst running with the extcon driver the value in the
arizona structure remains at zero and is treated as a short preventing
headphone output.

Change-Id: Id303054d1fa00483291ea9a6d35c75be190ceea7
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
(cherry picked from commit bdb62f5ac93c9cd22771f70672037d178e08e3ae)