ASoC: arizona: Correct relationship between VCO corner and Fref
When configuring the FLL we must ensure that the reference clock passed
to the FLL is under a certain limit. This limit was specified
incorrectly in the current code, this patch corrects this. Although the
error will only be encountered in some edge cases.

Change-Id: I8fe58e97b9e02dba62416429ce684e54dae56e29
Reported-by: Ryo Tsutsui <ryo.tsutsui@wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
(cherry picked from commit a0780c3833e6c37d19c75c2207b7aa2131c33121)