misc: mmi-factory: Handle GPIO defer
Add support to handle case where a GPIO chip controller may not yet be
ready when we are probing this device. Add handling to ensure the call
to of_get_gpio_flags succeeded, and if it hasn't, propagate the error up
to the caller so we can handle properly the -EPROBE_DEFER cases.

Change-Id: I9ef597d1614fc92e848d37b4f1bf68f10029ebde
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>