arm/dt: 8994: camera: Update IMX230 pwdn gpio
Updated the gpio configuration for the camera power down
gpio so that when the common file is included in the
portable, the gpio will not be included since it's not
part of the standard Flex.

It is needed for both bunkerhill boards since it is used
to enable other regulators on a daughter card board
that plugs into the generic CSI port. The gpio did
change between M1 and M2 from gpio 131 to gpio 126.

Change-Id: I582004a3185d6ea1acc7be8ebe0d054731f85a7b
Signed-off-by: Elena Satraitis <w04888@motorola.com>
Reviewed-on: http://gerrit.mot.com/681597
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Alice Zhang <w18656@motorola.com>
Reviewed-by: Dave Schmidt <d.schmidt@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>