arm/dt: 8994: Add wlan bootstrap gpio
GPIO 112 is also connected to the wifi IC and controls the booting
process. Specify this for the driver so it can properly control the
level during initial boot. QC only specifis this on their
msm8994-v2.dtsi, but we need it on our older v1 parts as wel so we'll
just define it across all 8994 products.

Change-Id: I984f26241919e613296f295358a36ec8930627f8
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>