ASoC: msm: florida: Add DSP 4 Virtual Output
Add a Wolfson DSP core 4 virtual output in the florida codec driver and
mark it as a DAPM ignore-suspend control in the ASoC msm machine driver.

A virtual output for the Wolfson codec's DSP core 4 is needed to allow a
complete end-to-end DAPM route so that the firmware and configuration files
will be loaded when the DSP core 4 is configured by the proper mixer
control settings.

Change-Id: If1e1514ad384022947660c1af87bdbf02297bc46
Signed-off-by: John Ratke (w16823) <w16823@motorola.com>
Reviewed-on: http://gerrit.mot.com/731389
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Christopher Fries <cfries@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>